FIx printing for Run and Qemu
